requirements.txt
langchain>=0.3.0
langchain-nvidia-ai-endpoints>=0.3.0
langgraph>=0.2.0
pinecone>=5.0.0
python-dotenv>=1.0.0
flask>=3.0.0
ipykernel>=6.0.0
notebook>=7.0.0
